Updated documentation, warnings, renamed  to ; reorganized util tests into one file
